tmux.conf: remove XT from terminal-overrides
From terminfo.src SCREEN Extensions:
    XT   (bool)  Terminal understands special xterm
                 sequences (OSC,  mouse tracking).
http://invisible-island.net/ncurses/terminfo.src.html#toc-_S_C_R_E_E_N__Extensions_

The OSC sequeces are used to set the title string in X.

This user-defined capabilty is already set in the terminfo
database for the relevant xterm flavours. It can be verified
by issuing the following command in a plain xterm.

```
    infocmp -x | grep ' XT'
```

Furthermore, to check if tmux has correclty indentified that
capabilty the following command can be used.

```
    tmux info | grep XT
```