tmux.conf: switch default-terminal to tmux-256colors
A dedicated entry for tmux has been added to the terminfo
database on 2015-05-02.

http://invisible-island.net/ncurses/terminfo.src.html#t20150502

It turned out that the $TERMCAP check broke since some vim
version and that the title setting actually works best with
t_ts and t_fs unset (that's basically how it was configured
when default-terminal was set to screen-256color).